fixed wall wextra werror

+++ source/ft_strlen_s.c
@@ -6,14 +6,14 @@
 /*   By: dnebatz <dnebatz@student.42wolfsburg.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/17 10:06:25 by dnebatz           #+#    #+#             */
-/*   Updated: 2023/07/25 21:40:51 by dnebatz          ###   ########.fr       */
+/*   Updated: 2023/08/03 17:22:41 by dnebatz          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/so_long.h"
 
 //return size until 0 or \n
-size_t	ft_strlen_s(const char *s)
+int	ft_strlen_s(const char *s)
 {
 	unsigned long	i;
